Add pkg/db package with Tx interface and refactor pkg/db/mongo package with new Tx interface
