+++ server/pkg/db/tx.go
@@ -1,0 +1,10 @@
+package db
+
+import "context"
+
+type TxFunc func(ctx context.Context) (interface{}, error)
+
+type Tx interface {
+	Execute(ctx context.Context, fn TxFunc) (interface{}, error)
+	Client() interface{}
+}
